requirements.txt
numpy
scipy
torch==2.0.1
torchvision
pytorch-lightning==1.9.5
pyarrow
pandas

pytest
tqdm

seaborn
opencv-python

black==24.1.1

geometry-score @ git+https://github.com/KhrulkovV/geometry-score
msid @ git+https://github.com/xgfs/imd

torch_geometric
ogb

six
tensorboard
matplotlib
scikit-image
scikit-learn
pre-commit>=2.15.0
# randaugment==1.0.2
gitpython
xmltodict
torchmetrics
imageio>=2.9.0
PyYaml>=5.4.1
pyDeprecate
medmnist
chardet
setuptools
Pillow
plotly
loguru
hug
huggingface_hub
jsonlines
textattack
sentence_transformers
hydra-core
ipykernel
huggingface-hub
ipywidgets
pyarrow
tokenizers
datasets
transformers
Cython
gudhi
evaluate
accelerate
albumentations
torch_geometric
torcheval
ogb
openpyxl
albumentations
einops
lightning-bolts
